remove 'delete' section from save_if_valid function

+++ submit_scripts/cutlass_utils.py
@@ -10,15 +10,29 @@ def save_if_valid(data_dict, nodetype):
             data_id = data_dict.id
             print("Succesfully saved {}. ID: {}".format(nodename,data_id))
             print(data_dict.to_json(indent=4))
-
-            deletion_success = data_dict.delete()
-            if deletion_success:
-                print("Deleted {} with ID {}".format(nodename,data_id))
-            else:
-                print("Deletion of {} with ID {} failed.".format(
-                    nodetype,data_id))
         else:
             print("Save failed")
+    else:
+        print("Invalid...")
+        validation_errors = data_dict.validate()
+        pprint(validation_errors)
+
+
+def delete_node(data_dict, nodetype):
+    # TODO: test delete_node() function
+    """Usage: delete_node( metadata_dict, NodeType(e.g.Project))"""
+    nodename = nodetype.__name__
+    if data_dict.is_valid():
+        print("Valid!")
+        success = data_dict.delete()
+
+        if success:
+            data_id = data_dict.id
+            print("Deleted {} with ID {}".format(nodename,data_id))
+            print(data_dict.to_json(indent=4))
+        else:
+            print("Deletion of {} with ID {} failed.".format(
+                    nodetype,data_id))
     else:
         print("Invalid...")
         validation_errors = data_dict.validate()
